pyproject.toml
[tool.poetry]
name = "project-12-django"
version = "0.1.0"
description = ""
authors = ["Xalva <maasania@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.0.1"
SQLite4 = "^0.1.1"
Jinja2 = "^3.1.3"
django-debug-toolbar = "^4.3.0"
pydantic = "^2.6.4"
django-filter = "^24.1"
celery = "^5.3.6"
redis = "^5.0.3"
django-environ = "^0.11.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
